[TASK] Consider parsed files on recompilation
If LESS files are using the @import statement, modifications
to the imported files are not considered and thus the specific
LESS file won't be recompiled.

The result of lessc is stored in the cache_hash cache which
also contains all parsed files that need to be checked of
future pre-compilation checks.

Compatibility: 4.5, 4.6, 4.7, 6.0, 6.1